+++ bigtree/__init__.py
@@ -1,5 +1,7 @@
 __version__ = "0.2.0"
 
+from bigtree.dag.construct import dataframe_to_dag, dict_to_dag, list_to_dag
+from bigtree.dag.export import dag_to_dot
 from bigtree.node.basenode import BaseNode
 from bigtree.node.dagnode import DAGNode
 from bigtree.node.node import Node
@@ -15,7 +17,6 @@ from bigtree.tree.construct import (
     nested_dict_to_tree,
 )
 from bigtree.tree.export import (
-    dag_to_dot,
     print_tree,
     tree_to_dataframe,
     tree_to_dict,
@@ -51,6 +52,7 @@ from bigtree.utils.exceptions import (
     TreeError,
 )
 from bigtree.utils.iterators import (
+    dag_iterator,
     levelorder_iter,
     levelordergroup_iter,
     postorder_iter,