+++ tests/node/test_dagnode.py
@@ -1,11 +1,13 @@
 import copy
 import unittest
 
+import pandas as pd
 import pytest
 
 from bigtree.node.dagnode import DAGNode
 from bigtree.node.node import Node
 from bigtree.utils.exceptions import LoopError, TreeError
+from bigtree.utils.iterators import dag_iterator
 from tests.conftest import assert_print_statement
 
 
@@ -29,6 +31,7 @@ class TestDAGNode(unittest.TestCase):
         self.e = DAGNode(name="e", age=35)
         self.f = DAGNode(name="f", age=38)
         self.g = DAGNode(name="g", age=10)
+        self.h = DAGNode(name="h", age=6)
 
     def tearDown(self):
         self.a = None
@@ -38,6 +41,7 @@ class TestDAGNode(unittest.TestCase):
         self.e = None
         self.f = None
         self.g = None
+        self.h = None
 
     def test_from_dict(self):
         self.a = DAGNode().from_dict({"name": "a", "age": 90})
@@ -49,108 +53,119 @@ class TestDAGNode(unittest.TestCase):
         self.g = DAGNode().from_dict({"name": "g", "age": 10})
         self.h = DAGNode().from_dict({"name": "h", "age": 6})
 
-        self.b.parents = [self.a]
-        self.c.parents = [self.b]
-        self.d.parents = [self.b]
-        self.e.parents = [self.c, self.d]
-        self.f.parents = [self.c, self.e]
-        self.g.parents = [self.f]
+        self.c.parents = [self.a, self.b]
+        self.d.parents = [self.a, self.c]
+        self.e.parents = [self.d]
+        self.f.parents = [self.c, self.d]
+        self.g.parents = [self.c]
+        self.h.parents = [self.g]
 
-        assert_tree_structure_dag_self(self)
+        assert_dag_structure_self(self)
+        assert_dag_structure_root(self.a)
 
     def test_set_parent_error(self):
         with pytest.raises(ValueError) as exc_info:
-            self.b.parent = self.a
+            self.c.parent = self.a
         assert str(exc_info.value).startswith("Attempting to set `parent` attribute")
 
         with pytest.raises(ValueError) as exc_info:
-            self.b = DAGNode("b", parent=self.a)
+            self.c = DAGNode("b", parent=self.a)
         assert str(exc_info.value).startswith("Attempting to set `parent` attribute")
 
         with pytest.raises(ValueError) as exc_info:
-            self.b.parent
+            self.c.parent
         assert str(exc_info.value).startswith("Attempting to access `parent` attribute")
 
     def test_set_parents(self):
-        self.b.parents = [self.a]
-        self.c.parents = [self.b]
-        self.d.parents = [self.b]
-        self.e.parents = [self.c, self.d]
-        self.f.parents = [self.c, self.e]
-        self.g.parents = [self.f]
+        self.c.parents = [self.a, self.b]
+        self.d.parents = [self.a, self.c]
+        self.e.parents = [self.d]
+        self.f.parents = [self.c, self.d]
+        self.g.parents = [self.c]
+        self.h.parents = [self.g]
 
-        assert_tree_structure_dag_self(self)
+        assert_dag_structure_self(self)
+        assert_dag_structure_root(self.a)
 
     def test_set_parents_rshift(self):
-        self.a >> self.b
+        self.a >> self.c
+        self.a >> self.d
         self.b >> self.c
-        self.b >> self.d
-        self.c >> self.e
+        self.c >> self.d
         self.c >> self.f
-        self.e >> self.f
-        self.f >> self.g
+        self.c >> self.g
         self.d >> self.e
+        self.d >> self.f
+        self.g >> self.h
 
-        assert_tree_structure_dag_self(self)
+        assert_dag_structure_self(self)
+        assert_dag_structure_root(self.a)
 
     def test_set_parents_lshift(self):
-        self.b << self.a
+        self.c << self.a
+        self.d << self.a
         self.c << self.b
-        self.d << self.b
-        self.e << self.c
-        self.e << self.d
+        self.d << self.c
         self.f << self.c
-        self.f << self.e
-        self.g << self.f
+        self.g << self.c
+        self.e << self.d
+        self.g << self.d
+        self.h << self.g
 
     def test_set_parents_constructor(self):
         self.a = DAGNode(name="a", age=90)
-        self.b = DAGNode(name="b", age=65, parents=[self.a])
-        self.c = DAGNode(name="c", age=60, parents=[self.b])
-        self.d = DAGNode(name="d", age=40, parents=[self.b])
-        self.e = DAGNode(name="e", age=35, parents=[self.c, self.d])
-        self.f = DAGNode(name="f", age=38, parents=[self.c, self.e])
-        self.g = DAGNode(name="g", age=10, parents=[self.f])
+        self.b = DAGNode(name="b", age=65)
+        self.c = DAGNode(name="c", age=60, parents=[self.a, self.b])
+        self.d = DAGNode(name="d", age=40, parents=[self.a, self.c])
+        self.e = DAGNode(name="e", age=35, parents=[self.d])
+        self.f = DAGNode(name="f", age=38, parents=[self.c, self.d])
+        self.g = DAGNode(name="g", age=10, parents=[self.c])
+        self.h = DAGNode(name="h", age=6, parents=[self.g])
 
-        assert_tree_structure_dag_self(self)
+        assert_dag_structure_self(self)
+        assert_dag_structure_root(self.a)
 
     def test_set_parents_none_parent(self):
-        self.b.parents = [self.a]
+        self.c.parents = [self.a]
         with pytest.raises(TypeError):
-            self.b.parents = None
+            self.c.parents = None
 
     def test_set_parent_duplicate(self):
         # Set parent again
-        self.b.parents = [self.a]
-        self.b.parents = [self.a]
+        self.c.parents = [self.a]
+        self.c.parents = [self.a]
 
     def test_set_parent_duplicate_constructor(self):
         # Set parent again
         self.a = DAGNode(name="a", age=90)
-        self.b = DAGNode(name="b", age=65, parents=[self.a])
-        self.b.parents = [self.a]
+        self.c = DAGNode(name="c", age=60, parents=[self.a])
+        self.c.parents = [self.a]
 
     def test_set_children(self):
-        self.a.children = [self.b]
-        self.b.children = [self.c, self.d]
-        self.c.children = [self.e, self.f]
-        self.d.children = [self.e]
-        self.e.children = [self.f]
-        self.f.children = [self.g]
-        self.g.children = []
+        self.a.children = [self.c, self.d]
+        self.b.children = [self.c]
+        self.c.children = [self.d, self.f, self.g]
+        self.d.children = [self.e, self.f]
+        self.g.children = [self.h]
 
-        assert_tree_structure_dag_self(self)
+        assert_dag_structure_self(self)
+        assert_dag_structure_root(self.a)
 
     def test_set_children_constructor(self):
-        self.g = DAGNode(name="g", age=10, children=[])
-        self.f = DAGNode(name="f", age=38, children=[self.g])
-        self.e = DAGNode(name="e", age=35, children=[self.f])
-        self.d = DAGNode(name="d", age=40, children=[self.e])
-        self.c = DAGNode(name="c", age=60, children=[self.e, self.f])
-        self.b = DAGNode(name="b", age=65, children=[self.c, self.d])
-        self.a = DAGNode(name="a", age=90, children=[self.b])
+        self.h = DAGNode(name="h", age=6)
+        self.g = DAGNode(name="g", age=10, children=[self.h])
+        self.f = DAGNode(name="f", age=38)
+        self.e = DAGNode(name="e", age=35)
+        self.d = DAGNode(name="d", age=40)
+        self.c = DAGNode(name="c", age=60)
+        self.b = DAGNode(name="b", age=65)
+        self.a = DAGNode(name="a", age=90, children=[self.c, self.d])
+        self.c.children = [self.d, self.f, self.g]
+        self.b.children = [self.c]
+        self.d.children = [self.e, self.f]
 
-        assert_tree_structure_dag_self(self)
+        assert_dag_structure_self(self)
+        assert_dag_structure_root(self.a)
 
     def test_set_children_none_children(self):
         with pytest.raises(TypeError):
@@ -158,44 +173,41 @@ class TestDAGNode(unittest.TestCase):
 
     def test_set_children_duplicate(self):
         # Set child again
-        self.a.children = [self.b]
-        self.a.children = [self.b]
+        self.a.children = [self.c]
+        self.a.children = [self.c]
 
     def test_set_children_duplicate_constructor(self):
         # Set child again
-        self.a = DAGNode(children=[self.b])
-        self.a.children = [self.b]
+        self.a = DAGNode(children=[self.c])
+        self.a.children = [self.c]
 
     def test_deep_copy_set_children(self):
-        self.b.parents = [self.a]
-        self.c.parents = [self.b]
-        self.d.parents = [self.b]
-        self.e.parents = [self.c, self.d]
-        self.f.parents = [self.c, self.e]
-        self.g.parents = [self.f]
+        self.c.parents = [self.a, self.b]
+        self.d.parents = [self.a, self.c]
+        self.e.parents = [self.d]
+        self.f.parents = [self.c, self.d]
+        self.g.parents = [self.c]
+        self.h.parents = [self.g]
 
         a2 = self.a.copy()
         assert (
-            len(list(a2.children)) == 1
+            len(list(a2.children)) == 2
         ), f"Expected 2 children, received {len(a2.children)}"
 
     def test_shallow_copy_set_children(self):
-        self.a.children = [self.b]
-        self.b.children = [self.c, self.d]
-        self.c.children = [self.e, self.f]
-        self.d.children = [self.e]
-        self.e.children = [self.f]
-        self.f.children = [self.g]
-        self.g.children = []
+        self.a.children = [self.c, self.d]
+        self.b.children = [self.c]
+        self.c.children = [self.d, self.f, self.g]
+        self.d.children = [self.e, self.f]
+        self.g.children = [self.h]
 
         a2 = copy.copy(self.a)
         assert (
-            len(list(a2.children)) == 1
-        ), f"Expected 1 child, received {len(list(a2.children))}"
+            len(list(a2.children)) == 2
+        ), f"Expected 2 child, received {len(list(a2.children))}"
         assert (
-            a2.children[0] == self.b or a2.children[1] == self.b
+            a2.children[0] == self.c or a2.children[1] == self.c
         ), "Shallow copy does not copy child nodes"
-        assert len(list(a2.children)) == 1, "Shallow copy does not copy child nodes"
 
     def test_error_set_parent_type_error(self):
         # Error: wrong type
@@ -263,29 +275,30 @@ class TestDAGNode(unittest.TestCase):
         ), f"Node description should be {expected}, but it is {actual}"
 
 
-def assert_tree_structure_dag_self(self):
+def assert_dag_structure_self(self):
     """
     Tree should have structure
-    a >> b
+    a >> c
+    a >> d
     b >> c
-    b >> d
-    c >> e
+    c >> d
     c >> f
-    e >> f
-    f >> g
+    c >> g
     d >> e
+    g >> h
     """
-    nodes = [self.a, self.b, self.c, self.d, self.e, self.f, self.g]
+    nodes = [self.a, self.b, self.c, self.d, self.e, self.f, self.g, self.h]
 
     # Test parent
     expected_ans = [
         [],
-        [self.a],
-        [self.b],
-        [self.b],
+        [],
+        [self.a, self.b],
+        [self.a, self.c],
+        [self.d],
         [self.c, self.d],
-        [self.c, self.e],
-        [self.f],
+        [self.c],
+        [self.g],
     ]
     for node, expected in zip(nodes, expected_ans):
         actual = list(node.parents)
@@ -295,12 +308,13 @@ def assert_tree_structure_dag_self(self):
 
     # Test children
     expected_ans = [
-        [self.b],
         [self.c, self.d],
+        [self.c],
+        [self.d, self.f, self.g],
         [self.e, self.f],
-        [self.e],
-        [self.f],
-        [self.g],
+        [],
+        [],
+        [self.h],
         [],
     ]
     for node, expected in zip(nodes, expected_ans):
@@ -312,12 +326,13 @@ def assert_tree_structure_dag_self(self):
     # Test ancestors
     expected_ans = [
         [],
-        [self.a],
+        [],
         [self.a, self.b],
-        [self.a, self.b],
+        [self.a, self.b, self.c],
         [self.a, self.b, self.c, self.d],
-        [self.a, self.b, self.c, self.d, self.e],
-        [self.a, self.b, self.c, self.d, self.e, self.f],
+        [self.a, self.b, self.c, self.d],
+        [self.a, self.b, self.c],
+        [self.a, self.b, self.c, self.g],
     ]
     for node, expected in zip(nodes, expected_ans):
         actual = list(node.ancestors)
@@ -327,12 +342,13 @@ def assert_tree_structure_dag_self(self):
 
     # Test descendants
     expected_ans = [
-        [self.b, self.c, self.e, self.f, self.g, self.d],
-        [self.c, self.e, self.f, self.g, self.d],
-        [self.e, self.f, self.g],
-        [self.e, self.f, self.g],
-        [self.f, self.g],
-        [self.g],
+        [self.c, self.d, self.e, self.f, self.g, self.h],
+        [self.c, self.d, self.e, self.f, self.g, self.h],
+        [self.d, self.e, self.f, self.g, self.h],
+        [self.e, self.f],
+        [],
+        [],
+        [self.h],
         [],
     ]
     for node, expected in zip(nodes, expected_ans):
@@ -346,9 +362,10 @@ def assert_tree_structure_dag_self(self):
         (),
         (),
         (self.d,),
-        (self.c,),
+        (self.c, self.f, self.g),
         (self.f,),
-        (self.e,),
+        (self.d, self.g, self.e),
+        (self.d, self.f),
         (),
     ]
     for node, expected in zip(nodes, expected_ans):
@@ -358,7 +375,7 @@ def assert_tree_structure_dag_self(self):
         ), f"Siblings of {node} is wrong, expected {expected}, received {actual}"
 
     # Test node_name
-    expected_ans = ["a", "b", "c", "d", "e", "f", "g"]
+    expected_ans = ["a", "b", "c", "d", "e", "f", "g", "h"]
     for node, expected in zip(nodes, expected_ans):
         actual = node.node_name
         assert (
@@ -366,7 +383,7 @@ def assert_tree_structure_dag_self(self):
         ), f"node_name of {node} is wrong, expected {expected}, received {actual}"
 
     # Test is_root
-    expected_ans = [True, False, False, False, False, False, False]
+    expected_ans = [True, True, False, False, False, False, False, False]
     for node, expected in zip(nodes, expected_ans):
         actual = node.is_root
         assert (
@@ -374,7 +391,7 @@ def assert_tree_structure_dag_self(self):
         ), f"is_root of {node} is wrong, expected {expected}, received {actual}"
 
     # Test is_leaf
-    expected_ans = [False, False, False, False, False, False, True]
+    expected_ans = [False, False, False, False, True, True, False, True]
     for node, expected in zip(nodes, expected_ans):
         actual = node.is_leaf
         assert (
@@ -422,15 +439,69 @@ def assert_tree_structure_dag_self(self):
         "DAGNode(e, age=35)\n",
         "DAGNode(f, age=38)\n",
         "DAGNode(g, age=10)\n",
+        "DAGNode(h, age=6)\n",
     ]
     for node, expected in zip(nodes, expected_ans):
         assert_print_statement(print, expected, node)
 
     # Test set_attrs()
-    attrs = [1, 2, 3, 4, 5, 6, 7]
+    attrs = [1, 2, 3, 4, 5, 6, 7, 8]
     for node, expected in zip(nodes, attrs):
         node.set_attrs({"index": expected})
         actual = node.get_attr("index")
         assert (
             actual == expected
         ), f"Node attribute should be {expected}, but it is {actual}"
+
+
+def assert_dag_structure_root(dag):
+    expected = [
+        ("a", "c"),
+        ("a", "d"),
+        ("b", "c"),
+        ("c", "d"),
+        ("c", "f"),
+        ("c", "g"),
+        ("d", "e"),
+        ("d", "f"),
+        ("g", "h"),
+    ]
+    actual = [
+        (parent.node_name, child.node_name) for parent, child in dag_iterator(dag)
+    ]
+    len_expected = 9
+    len_actual = len(actual)
+    for relation in actual:
+        if relation[1] in ["a", "b"]:
+            len_expected = 11
+            assert pd.isnull(relation[0]), f"Expected\n{relation}\nReceived\n{actual}"
+        else:
+            assert relation in expected, f"Expected\n{relation}\nReceived\n{actual}"
+    assert (
+        len_expected == len_actual
+    ), f"Expected\n{len_expected}\nReceived\n{len_actual}"
+
+
+def assert_dag_structure_attr_root(dag):
+    expected_dict = {
+        "a": 90,
+        "b": 65,
+        "c": 60,
+        "d": 40,
+        "e": 35,
+        "f": 38,
+        "g": 10,
+        "h": 6,
+    }
+    for parent, child in dag_iterator(dag):
+        expected = expected_dict[parent.node_name]
+        actual = parent.age
+        assert (
+            expected == actual
+        ), f"For {parent}, expected\n{expected}\nReceived\n{actual}"
+
+        expected = expected_dict[child.node_name]
+        actual = child.age
+        assert (
+            expected == actual
+        ), f"For {child}, expected\n{expected}\nReceived\n{actual}"