+++ CHANGELOG.md
@@ -7,7 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Work In Progress
 - Node: WeightedNode for weighted edge tree implementation.
-- Tree Exporter: More customizations for Node to dot and DAGNode to dot.
+
+## [0.3.0]
+### Added
+- DAG Constructors: From list, nested dictionary, pandas DataFrame.
+- Tree Exporter and DAG Exporter: More customizations for Node to dot and DAGNode to dot.
+- Utility Iterators: DAG traversal methods.
 
 ## [0.2.0] - 2022-11-03
 ### Added
@@ -27,5 +32,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Workflow To Do App: Tree use case with to-do list implementation.
 
 [Unreleased]: https://github.com/kayjan/bigtree/compare/HEAD...v0.2.0
+[0.3.0]: https://github.com/kayjan/bigtree/compare/v0.2.0...v0.3.0
 [0.2.0]: https://github.com/kayjan/bigtree/compare/v0.1.0...v0.2.0
 [0.1.0]: https://github.com/kayjan/bigtree/releases/tag/v0.1.0